wasserstoff/AiTask/requirements.txt
requests
sentence-transformers
faiss-cpu
flask
numpy
